+++ MessageLogger/Models/Message.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MessageLogger
+namespace MessageLogger.Models
 {
     public class Message
     {
